[Refactoring] : File Constructure
