+++ gui/ai_panel.py
@@ -1,9 +1,9 @@
-from gui.componets import CodeBox
 import wx
-from gui.componets.SVGButton import SVGButton
+from gui.componets import CodeBox
+from gui.componets.common.SVGButton import SVGButton
 from gui.componets.PromptInputPanel import PromptInputPanel
-from gui.componets.AIChatBox import AIChatBox
-from gui.componets.MyChatBox import MyChatBox
+from gui.componets.chat.AIChatBox import AIChatBox
+from gui.componets.chat.MyChatBox import MyChatBox
 from utils.db_handler import create_conversation, get_code_blocks, get_messages
