+++ gui/componets/chat/AIChatBox.py
@@ -1,5 +1,5 @@
 import wx
-from .SVGButton import SVGButton
+from ..common.SVGButton import SVGButton
 
 
 class AIChatBox(wx.Panel):
@@ -62,7 +62,7 @@ class AIChatBox(wx.Panel):
                                           wx.FONTSTYLE_NORMAL, wx.FONTWEIGHT_NORMAL))
             text_sizer.Add(message_label, 0, wx.ALL | wx.EXPAND, 3)
         else:
-            from .CodeBox import CodeBox
+            from ..CodeBox import CodeBox
             # 코드일 경우 CodeBox 사용
             code_box = CodeBox(
                 self, isWorkflow=False, texts=message, language=type)
